Refactor: Split Home page into components
Split the Home page into smaller, reusable components for better organization and readability.
