+++ src/components/sections/FeaturedProjects.tsx
@@ -1,0 +1,39 @@
+
+import { Button } from "@/components/ui/button";
+import { Link } from "react-router-dom";
+import ProjectCard from "@/components/ProjectCard";
+import { projects } from "@/data/projects";
+
+const FeaturedProjects = () => {
+  const featuredProjects = projects.slice(0, 3);
+
+  return (
+    <section className="py-16">
+      <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8">
+        <div className="flex justify-between items-center mb-12">
+          <h2 className="section-heading mb-0">Featured Projects</h2>
+          <Button asChild variant="outline">
+            <Link to="/projects">View All Projects</Link>
+          </Button>
+        </div>
+
+        <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-8">
+          {featuredProjects.map((project) => (
+            <ProjectCard
+              key={project.id}
+              id={project.id}
+              title={project.title}
+              description={project.description}
+              tags={project.tags}
+              image={project.image}
+              githubUrl={project.githubUrl}
+              liveUrl={project.liveUrl}
+            />
+          ))}
+        </div>
+      </div>
+    </section>
+  );
+};
+
+export default FeaturedProjects;
